Inital material changes
Issue: https://github.com/adonisv79/bytecommander.com/issues/3
Added material changes for react. so far we have played with some buttons and grids which are pretty similar to bootstrap. We will be playing with styles and themes better in the next push as well as fix those disalligned UI parts.