Majority ~80% of AboutUS (CV Profile) is completed
issue: https://github.com/adonisv79/bytecommander.com/issues/2
this covers most of the main page revamp which fixes a lot of the old idle static code. we have finished the accordion view and have tested the UI orientations to work well on small devices. New accordion module is created to make it easier to use this collapsible component in the future.