Broken down code to create a reusable Game Engine
The previous version was made to show off a quick game loop systems in canvas but it is not meant to be scalable and reused. We now have broken up the parts to create our own Game Engine for the web. this will be removed from this to spin off a new product